Make the GET DIAGNOSTICS counter 64 bit safe
Along the way update all places which work with the
number of tuples, one way or another.

Cherry-pick c970292a94e0fff468d500db430d751b6221a0b4 from upstream

+++ src/backend/cdb/cdbresynchronizechangetracking.c
@@ -1312,7 +1312,7 @@ ChangeTracking_GetIncrementalChangeList(void)
 	IncrementalChangeList* result = NULL;
 	StringInfoData 	sqlstmt;
 	int 			ret;
-	int 			proc;
+	int 			proc;	/* 32 bit, only holds metadata */
 	volatile bool 	connected = false;
 	ResourceOwner 	save = CurrentResourceOwner;
 	MemoryContext 	oldcontext = CurrentMemoryContext;
@@ -1366,7 +1366,7 @@ ChangeTracking_GetIncrementalChangeList(void)
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, true, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 
 		if (ret > 0 && SPI_tuptable != NULL)
@@ -1580,7 +1580,7 @@ ChangeTrackingResult* ChangeTracking_GetChanges(ChangeTrackingRequest *request)
 	ChangeTrackingResult* result = NULL;
 	StringInfoData 	sqlstmt;
 	int 			ret;
-	int 			proc;
+	int 			proc;	/* 32 bit, only holds metadata */
 	int				i;
 	volatile bool 	connected = false; /* needs to survive PG_TRY()/CATCH() */
 	ResourceOwner 	save = CurrentResourceOwner;
@@ -1652,7 +1652,7 @@ ChangeTrackingResult* ChangeTracking_GetChanges(ChangeTrackingRequest *request)
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, true, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 
 		if (ret > 0 && SPI_tuptable != NULL)
@@ -3077,7 +3077,7 @@ int ChangeTracking_CompactLogFile(CTFType source, CTFType dest, XLogRecPtr*	upto
 {	
 	StringInfoData 	sqlstmt;
 	int 			ret;
-	int 			proc;
+	int 			proc;	/* 32 bit, only holds metadata */
 	bool 			connected = false;
 	int64			count = 0;
 	ResourceOwner 	save = CurrentResourceOwner;
@@ -3121,7 +3121,7 @@ int ChangeTracking_CompactLogFile(CTFType source, CTFType dest, XLogRecPtr*	upto
 
 		/* Do the query. */
 		ret = SPI_execute(sqlstmt.data, true, 0);
-		proc = SPI_processed;
+		proc = (int) SPI_processed;
 
 		if ((segmentState != SegmentStateChangeTrackingDisabled) &&
 			ret > 0 && SPI_tuptable != NULL)
